+++ NEWS.md
@@ -1,3 +1,9 @@
+# Changes in version 0.1.9 (2026-03-31)
+
+* improved error messages returned by `wd_getdata_syn()` and `wd_getdata_wwg()`
+
+* added `wd_getdata_checks()`, an internal function used to check the arguments passed to `wd_getdata*()`
+
 # Changes in version 0.1.8 (2025-10-28)
 
 * modified `wd_getdata_syn()` to convert NULL values returned by the API to NAs
@@ -5,3 +11,4 @@
 # wthrdat 0.0.0.9000
 
 * Initial version
+